+++ packages/tendermint-rpc/src/client.spec.ts
@@ -6,7 +6,7 @@ import { ReadonlyDate } from "readonly-date";
 import { Stream } from "xstream";
 
 import { Adaptor } from "./adaptor";
-import { adaptorForVersion } from "./adaptorforversion";
+import { adaptorForVersion } from "./adaptors";
 import { Client } from "./client";
 import { ExpectedValues, tendermintInstances } from "./config.spec";
 import { buildQuery } from "./requests";
@@ -42,24 +42,26 @@ function randomString(): string {
 }
 
 function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expected: ExpectedValues): void {
-  it("can connect to tendermint with known version", async () => {
-    pendingWithoutTendermint();
-    const client = new Client(rpcFactory(), adaptor);
-    expect(await client.abciInfo()).toBeTruthy();
-    client.disconnect();
-  });
+  describe("create", () => {
+    it("can auto-discover Tendermint version and communicate", async () => {
+      pendingWithoutTendermint();
+      const client = await Client.create(rpcFactory());
+      const info = await client.abciInfo();
+      expect(info).toBeTruthy();
+      client.disconnect();
+    });
 
-  it("can auto-discover tendermint version and connect", async () => {
-    pendingWithoutTendermint();
-    const client = await Client.detectVersion(rpcFactory());
-    const info = await client.abciInfo();
-    expect(info).toBeTruthy();
-    client.disconnect();
+    it("can connect to Tendermint with known version", async () => {
+      pendingWithoutTendermint();
+      const client = await Client.create(rpcFactory(), adaptor);
+      expect(await client.abciInfo()).toBeTruthy();
+      client.disconnect();
+    });
   });
 
   it("can get genesis", async () => {
     pendingWithoutTendermint();
-    const client = new Client(rpcFactory(), adaptor);
+    const client = await Client.create(rpcFactory(), adaptor);
     const genesis = await client.genesis();
     expect(genesis).toBeTruthy();
     client.disconnect();
@@ -67,7 +69,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
 
   it("can broadcast a transaction", async () => {
     pendingWithoutTendermint();
-    const client = new Client(rpcFactory(), adaptor);
+    const client = await Client.create(rpcFactory(), adaptor);
     const tx = buildKvTx(randomString(), randomString());
 
     const response = await client.broadcastTxCommit({ tx: tx });
@@ -85,7 +87,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
 
   it("gets the same tx hash from backend as calculated locally", async () => {
     pendingWithoutTendermint();
-    const client = new Client(rpcFactory(), adaptor);
+    const client = await Client.create(rpcFactory(), adaptor);
     const tx = buildKvTx(randomString(), randomString());
     const calculatedTxHash = adaptor.hashTx(tx);
 
@@ -97,7 +99,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
 
   it("can query the state", async () => {
     pendingWithoutTendermint();
-    const client = new Client(rpcFactory(), adaptor);
+    const client = await Client.create(rpcFactory(), adaptor);
 
     const key = randomString();
     const value = randomString();
@@ -116,7 +118,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
 
   it("can call a bunch of methods", async () => {
     pendingWithoutTendermint();
-    const client = new Client(rpcFactory(), adaptor);
+    const client = await Client.create(rpcFactory(), adaptor);
 
     expect(await client.block()).toBeTruthy();
     expect(await client.commit(4)).toBeTruthy();
@@ -130,7 +132,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
   describe("status", () => {
     it("works", async () => {
       pendingWithoutTendermint();
-      const client = new Client(rpcFactory(), adaptor);
+      const client = await Client.create(rpcFactory(), adaptor);
 
       const status = await client.status();
 
@@ -159,7 +161,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
   describe("blockResults", () => {
     it("works", async () => {
       pendingWithoutTendermint();
-      const client = new Client(rpcFactory(), adaptor);
+      const client = await Client.create(rpcFactory(), adaptor);
 
       const height = 3;
       const results = await client.blockResults(height);
@@ -175,7 +177,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
   describe("blockchain", () => {
     it("returns latest in descending order by default", async () => {
       pendingWithoutTendermint();
-      const client = new Client(rpcFactory(), adaptor);
+      const client = await Client.create(rpcFactory(), adaptor);
 
       // Run in parallel to increase chance there is no block between the calls
       const [status, blockchain] = await Promise.all([client.status(), client.blockchain()]);
@@ -192,7 +194,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
 
     it("can limit by maxHeight", async () => {
       pendingWithoutTendermint();
-      const client = new Client(rpcFactory(), adaptor);
+      const client = await Client.create(rpcFactory(), adaptor);
 
       const height = (await client.status()).syncInfo.latestBlockHeight;
       const blockchain = await client.blockchain(undefined, height - 1);
@@ -206,7 +208,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
 
     it("works with maxHeight in the future", async () => {
       pendingWithoutTendermint();
-      const client = new Client(rpcFactory(), adaptor);
+      const client = await Client.create(rpcFactory(), adaptor);
 
       const height = (await client.status()).syncInfo.latestBlockHeight;
       const blockchain = await client.blockchain(undefined, height + 20);
@@ -220,7 +222,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
 
     it("can limit by minHeight and maxHeight", async () => {
       pendingWithoutTendermint();
-      const client = new Client(rpcFactory(), adaptor);
+      const client = await Client.create(rpcFactory(), adaptor);
 
       const height = (await client.status()).syncInfo.latestBlockHeight;
       const blockchain = await client.blockchain(height - 2, height - 1);
@@ -234,7 +236,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
 
     it("contains all the info", async () => {
       pendingWithoutTendermint();
-      const client = new Client(rpcFactory(), adaptor);
+      const client = await Client.create(rpcFactory(), adaptor);
 
       const height = (await client.status()).syncInfo.latestBlockHeight;
       const blockchain = await client.blockchain(height - 1, height - 1);
@@ -264,7 +266,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
   describe("tx", () => {
     it("can query a tx properly", async () => {
       pendingWithoutTendermint();
-      const client = new Client(rpcFactory(), adaptor);
+      const client = await Client.create(rpcFactory(), adaptor);
 
       const find = randomString();
       const me = randomString();
@@ -321,7 +323,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
 
     beforeAll(async () => {
       if (tendermintEnabled()) {
-        const client = new Client(rpcFactory(), adaptor);
+        const client = await Client.create(rpcFactory(), adaptor);
 
         // eslint-disable-next-line no-inner-declarations
         async function sendTx(): Promise<void> {
@@ -347,7 +349,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
 
     it("can paginate over txSearch results", async () => {
       pendingWithoutTendermint();
-      const client = new Client(rpcFactory(), adaptor);
+      const client = await Client.create(rpcFactory(), adaptor);
 
       const query = buildQuery({ tags: [{ key: "app.key", value: key }] });
 
@@ -366,7 +368,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
 
     it("can get all search results in one call", async () => {
       pendingWithoutTendermint();
-      const client = new Client(rpcFactory(), adaptor);
+      const client = await Client.create(rpcFactory(), adaptor);
 
       const query = buildQuery({ tags: [{ key: "app.key", value: key }] });
 
@@ -391,7 +393,7 @@ function websocketTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expec
 
     (async () => {
       const events: responses.NewBlockHeaderEvent[] = [];
-      const client = new Client(rpcFactory(), adaptor);
+      const client = await Client.create(rpcFactory(), adaptor);
       const stream = client.subscribeNewBlockHeader();
       expect(stream).toBeTruthy();
       const subscription = stream.subscribe({
@@ -449,7 +451,7 @@ function websocketTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expec
     const transactionData2 = buildKvTx(randomString(), randomString());
 
     const events: responses.NewBlockEvent[] = [];
-    const client = new Client(rpcFactory(), adaptor);
+    const client = await Client.create(rpcFactory(), adaptor);
     const stream = client.subscribeNewBlock();
     const subscription = stream.subscribe({
       next: (event) => {
@@ -505,7 +507,7 @@ function websocketTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expec
     pendingWithoutTendermint();
 
     const events: responses.TxEvent[] = [];
-    const client = new Client(rpcFactory(), adaptor);
+    const client = await Client.create(rpcFactory(), adaptor);
     const stream = client.subscribeTx();
     const subscription = stream.subscribe({
       next: (event) => {
@@ -549,7 +551,7 @@ function websocketTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expec
     const transactionData2 = buildKvTx(randomString(), randomString());
 
     const events: responses.TxEvent[] = [];
-    const client = new Client(rpcFactory(), adaptor);
+    const client = await Client.create(rpcFactory(), adaptor);
     const query = buildQuery({ tags: [{ key: "app.creator", value: expected.appCreator }] });
     const stream = client.subscribeTx(query);
     expect(stream).toBeTruthy();
@@ -587,7 +589,7 @@ function websocketTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expec
   it("can unsubscribe and re-subscribe to the same stream", async () => {
     pendingWithoutTendermint();
 
-    const client = new Client(rpcFactory(), adaptor);
+    const client = await Client.create(rpcFactory(), adaptor);
     const stream = client.subscribeNewBlockHeader();
 
     const event1 = await firstEvent(stream);
@@ -620,7 +622,7 @@ function websocketTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expec
   it("can subscribe twice", async () => {
     pendingWithoutTendermint();
 
-    const client = new Client(rpcFactory(), adaptor);
+    const client = await Client.create(rpcFactory(), adaptor);
     const stream1 = client.subscribeNewBlockHeader();
     const stream2 = client.subscribeNewBlockHeader();