+++ packages/tendermint-rpc/types/adaptors/v0-33/index.d.ts
@@ -1,0 +1,2 @@
+import { Adaptor } from "../../adaptor";
+export declare const v0_33: Adaptor;